pyproject.toml
[tool.coverage.run]
omit = ["github_metrics/tests/*"]

[tool.coverage.report]
fail_under = 90
skip_empty = true

[tool.coverage.html]
directory = ".tests_coverage"

[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "PLC0415", "ARG", "RUF059"]

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true
strict_equality = true
extra_checks = true
warn_unused_configs = true
warn_redundant_casts = true

[project]
name = "github-metrics"
version = "0.1.0"
description = "Standalone metrics service for GitHub webhook event tracking and visualization"
readme = "README.md"
requires-python = ">=3.13"
license = "Apache-2.0"
authors = [{ name = "myakove" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: FastAPI",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "asyncpg>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "alembic>=1.14.0",
  "pydantic>=2.9.0",
  "python-simple-logger>=1.0.40",
  "PyGithub>=2.4.0",
  "httpx>=0.28.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ipdb>=0.13.13", "ipython>=8.12.3"]
tests = [
  "pytest-asyncio>=0.26.0",
  "pytest-xdist>=3.7.0",
  "pytest-cov>=6.0.0",
  "pytest-mock>=3.14.0",
  "pytest>=8.3.3",
]
